refactor: Move utility functions and resolve warnings
